move consul mock to mock dir
