+++ components/pkg/mock/consul_mock.go
@@ -16,7 +16,7 @@
 // Source: ./consul.go
 
 // Package utils is a generated GoMock package.
-package utils
+package mock
 
 import (
 	reflect "reflect"
